Revendor g/g 1.92.
This removes all the migration logic from CCD to GNA. We can actually start to use this with g/g 1.90 already, I guess.
