+++ pkg/controller/operatingsystemconfig/actuator_test.go
@@ -66,15 +66,12 @@ var _ = Describe("Actuator", func() {
 		When("purpose is 'provision'", func() {
 			Describe("#Reconcile", func() {
 				It("should not return an error", func() {
-					userData, command, unitNames, fileNames, extensionUnits, extensionFiles, err := actuator.Reconcile(ctx, log, osc)
+					userData, extensionUnits, extensionFiles, err := actuator.Reconcile(ctx, log, osc)
 					Expect(err).NotTo(HaveOccurred())
 
 					Expect(string(userData)).To(ContainSubstring("/etc/containerd/config.toml"))
 					Expect(string(userData)).To(HavePrefix("{")) // check we have ignition format
 					Expect(string(userData)).To(HaveSuffix("}")) // check we have ignition format
-					Expect(command).To(BeNil())
-					Expect(unitNames).To(ConsistOf("some-unit.service"))
-					Expect(fileNames).To(ConsistOf("/some/file"))
 					Expect(extensionUnits).To(BeEmpty())
 					Expect(extensionFiles).To(BeEmpty())
 				})
@@ -88,15 +85,150 @@ var _ = Describe("Actuator", func() {
 
 			Describe("#Reconcile", func() {
 				It("should not return an error", func() {
-					userData, command, unitNames, fileNames, _, _, err := actuator.Reconcile(ctx, log, osc)
+					userData, extensionUnits, extensionFiles, err := actuator.Reconcile(ctx, log, osc)
 					Expect(err).NotTo(HaveOccurred())
 
-					Expect(userData).NotTo(BeEmpty()) // legacy logic is tested in ./generator/generator_test.go
-					Expect(command).To(BeNil())
-					Expect(unitNames).To(ConsistOf("some-unit.service"))
-					Expect(fileNames).To(ConsistOf("/some/file"))
+					Expect(userData).To(BeEmpty())
+					Expect(extensionFiles).To(ConsistOf(
+						extensionsv1alpha1.File{
+							Path:        "/etc/containerd/config.toml",
+							Permissions: ptr.To(int32(0644)),
+							Content: extensionsv1alpha1.FileContent{Inline: &extensionsv1alpha1.FileContentInline{Data: `# Generated by os-extension-metal
+imports = ["/etc/containerd/conf.d/*.toml"]
+disabled_plugins = []
+
+[plugins."io.containerd.grpc.v1.cri".registry]
+  config_path = "/etc/containerd/certs.d"
+`}},
+						},
+						extensionsv1alpha1.File{
+							Path:        "/some/file",
+							Permissions: nil,
+							Content:     extensionsv1alpha1.FileContent{Inline: &extensionsv1alpha1.FileContentInline{Data: "bar"}},
+						},
+					))
+					Expect(extensionUnits).To(ContainElement(
+						extensionsv1alpha1.Unit{
+							Name:    "some-unit.service",
+							Content: ptr.To("foo"),
+						},
+					))
 				})
 			})
 		})
 	})
 })
+
+// TODO
+// {
+// 	name: "containerd with network isolation",
+// 	config: &extensionsv1alpha1.OperatingSystemConfig{
+// 		Object: &extensionsv1alpha1.OperatingSystemConfig{
+// 			Spec: extensionsv1alpha1.OperatingSystemConfigSpec{
+// 				CRI: &extensionsv1alpha1.CRIConfig{
+// 					Name: extensionsv1alpha1.CRINameContainerD,
+// 				},
+// 				DefaultSpec: extensionsv1alpha1.DefaultSpec{
+// 					ProviderConfig: &runtime.RawExtension{
+// 						Raw: mustMarshal(t, &metalextensionv1alpha1.ImageProviderConfig{
+// 							NetworkIsolation: &metalextensionv1alpha1.NetworkIsolation{
+// 								AllowedNetworks: metalextensionv1alpha1.AllowedNetworks{
+// 									Ingress: []string{"10.0.0.1/24"},
+// 									Egress:  []string{"100.0.0.1/24"},
+// 								},
+// 								DNSServers: []string{"1.1.1.1", "1.0.0.1"},
+// 								NTPServers: []string{"134.60.1.27", "134.60.111.110"},
+// 								RegistryMirrors: []metalextensionv1alpha1.RegistryMirror{
+// 									{
+// 										Name:     "metal-stack registry",
+// 										Endpoint: "https://r.metal-stack.dev",
+// 										IP:       "1.2.3.4",
+// 										Port:     443,
+// 										MirrorOf: []string{
+// 											"ghcr.io",
+// 											"quay.io",
+// 										},
+// 									},
+// 									{
+// 										Name:     "local registry",
+// 										Endpoint: "http://localhost:8080",
+// 										IP:       "127.0.0.1",
+// 										Port:     8080,
+// 										MirrorOf: []string{
+// 											"docker.io",
+// 										},
+// 									},
+// 								},
+// 							}}),
+// 					},
+// 				},
+// 			},
+// 		},
+// 	},
+// 	wantErr: false,
+// 	want: types.Config{
+// 		Storage: types.Storage{
+// 			Files: []types.File{
+// 				{
+// 					Path: "/etc/systemd/resolved.conf.d/dns.conf",
+// 					Contents: types.FileContents{
+// 						Inline: `# Generated by os-extension-metal
+
+// [Resolve]
+// DNS=1.1.1.1 1.0.0.1
+// Domain=~.
+
+// `,
+// 					},
+// 				},
+// 				{
+// 					Path: "/etc/resolv.conf",
+// 					Contents: types.FileContents{
+// 						Inline: `# Generated by os-extension-metal
+// nameserver 1.1.1.1
+// nameserver 1.0.0.1
+// `,
+// 					},
+// 				},
+// 				{
+// 					Path: "/etc/systemd/timesyncd.conf",
+// 					Contents: types.FileContents{
+// 						Inline: `# Generated by os-extension-metal
+
+// [Time]
+// NTP=134.60.1.27 134.60.111.110
+// `,
+// 					},
+// 				},
+// 				{
+// 					Filesystem: "root",
+// 					Path:       "/etc/containerd/config.toml",
+// 					Mode:       &types.DefaultFileMode,
+// 					Contents: types.FileContents{
+// 						Inline: containerdConfig,
+// 					},
+// 					Overwrite: ptr.To(true),
+// 				},
+// 				{
+// 					Filesystem: "root",
+// 					Path:       "/etc/containerd/conf.d/isolated-cluster.toml",
+// 					Mode:       &types.DefaultFileMode,
+// 					Contents: types.FileContents{
+// 						Inline: `# Generated by os-extension-metal
+// version = 2
+
+// [plugins."io.containerd.grpc.v1.cri".registry]
+// [plugins."io.containerd.grpc.v1.cri".registry.mirrors]
+// [plugins."io.containerd.grpc.v1.cri".registry.mirrors."ghcr.io"]
+// endpoint = ["https://r.metal-stack.dev"]
+// [plugins."io.containerd.grpc.v1.cri".registry.mirrors."quay.io"]
+// endpoint = ["https://r.metal-stack.dev"]
+// [plugins."io.containerd.grpc.v1.cri".registry.mirrors."docker.io"]
+// endpoint = ["http://localhost:8080"]
+// `,
+// 					},
+// 				},
+// 			},
+// 		},
+// 	},
+// },
